#31 - VWorld api 서비스 구조 리팩토링
